# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Piggy-Edu API
  description: API for Piggy-Edu educational platform
  version: 1.0.1
servers:
  - url: https://piggy-edu-production.up.railway.app/
    description: Production server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required:
        - wallet
        - name
      properties:
        wallet:
          type: string
          example: "0x1234567890abcdef"
          description: User's blockchain wallet address
    
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: "Login successful"
        user:
          type: object
          properties:
            wallet:
              type: string
              example: "0x1234567890abcdef"
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    
    Level:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Nivel 1"
        description:
          type: string
          example: "Introducción a las finanzas"
        questions:
          type: array
          items:
            type: object
            description: Question objects array (structure to be defined)
    
    User:
      type: object
      properties:
        wallet:
          type: string
          example: "0x1234567890abcdef"
        level:
          type: integer
          example: 3
          description: User's current level in the system
        unmintedTokens:
          type: array
          description: Tokens earned but not yet minted
          items:
            $ref: '#/components/schemas/Token'
        mintedTokens:
          type: array
          description: Tokens already minted
          items:
            $ref: '#/components/schemas/Token'

    Token:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "NFT Nivel 1"
        address:
          type: string
          example: "0xabcdef1234567890"
    
    Paymaster:
      type: object
      properties:
        name:
          type: string
          example: "Paymaster address"
        address:
          type: string
          example: "0xDe6F1e5C07930AA9A65C9781A4885E6Bc76F9C52"
    
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Error message"

paths:
  /login:
    post:
      summary: User login
      description: Authenticate user with wallet address and name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /levels/{levelId}:
    get:
      summary: Get level data
      description: Retrieve information about a specific level
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: levelId
          required: true
          schema:
            type: integer
          description: ID of the level to retrieve
      responses:
        '200':
          description: Level data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Level'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Level not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /tokens/{tokenId}:
    get:
      summary: Get token address
      description: Retrieve information about a specific level NFT
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: tokenId
          required: true
          schema:
            type: integer
          description: ID of the token address to retrieve
      responses:
        '200':
          description: Token address retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Unauthorized - Invalid or missing authorization token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Token not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /user:
    get:
      summary: Get user data
      description: Retrieve current user's information
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /user/level/{levelId}:
    put:
      summary: Increment user level and add unminted token
      description: Increment the level of the current authenticated user by 1 and assign the unminted token for the specified level. If the user already has the token for the level, it will not be added again.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: levelId
          required: true
          schema:
            type: integer
          description: ID of the level to assign
      responses:
        '200':
          description: Level updated and token added successfully. If the token already exists, it is not added again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request - Invalid operation (e.g., trying to skip levels).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User or token not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /paymaster:
    get:
      summary: Get paymaster data
      description: Retrieve the paymaster name and address
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Paymaster data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paymaster'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error reading the paymaster data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'